server/requirements.txt
flask
flask-cors
flask-jwt-extended
flask_wtf
python-dotenv
mysql-connector-python